refactor: move downloadAllNftMediaForAccount to downloadNftMedia.ts
