+++ packages/shared/lib/core/nfts/utils/index.ts
@@ -3,7 +3,6 @@ export * from './composeUrlFromNftUri'
 export * from './convertAndFormatNftMetadata'
 export * from './getSpendableStatusFromUnspentNftOutput'
 export * from './downloadNftMedia'
-export * from './downloadAllNftMediaForAccount'
 export * from './fetchWithTimeout'
 export * from './parseNftMetadata'
 export * from './rewriteIpfsUri'
